+++ src/util.jl
@@ -191,9 +191,13 @@ end
 #---Minuit property accessors-----------------------------------------------------------------------
 function getproperty(m::Minuit, name::Symbol)
     if name == :values
-        return ValueView(m)
+        vs = ValueView(m)
+        ca_axes = m.componentarray_axes
+        return isnothing(ca_axes) ? vs : ComponentArray(collect(vs), ca_axes)
     elseif name == :errors
-        return ErrorView(m)
+        es = ErrorView(m)
+        ca_axes = m.componentarray_axes
+        return isnothing(ca_axes) ? es : ComponentArray(collect(es), ca_axes)
     elseif name == :fixed
         return FixedView(m)
     elseif name == :limits