+++ test/runtests.jl
@@ -1,8 +1,12 @@
-using Minuit2
+using ComponentArrays: ComponentArray
 using Minuit2.RooFit
+using Random: seed!
+using Minuit2
+using CxxWrap
 using Test
+include("functions.jl")
 
-@testset "Minuit2 tests" verbose = true begin 
+@testset "Minuit2 tests" verbose = true begin
     include("test_wrapper.jl")
     include("test_functions.jl")
     include("test_util.jl")
@@ -10,4 +14,5 @@ using Test
     include("test_cost.jl")
     include("test_roofit.jl")
     include("test_threads.jl")
+    include("test_optimization_integration.jl")
 end