+++ docs/src/release_notes.md
@@ -1,6 +1,11 @@
 
 # Release Notes
 
+## Latest
+### Additions
+- `m = Minuit(f, x)` can be set up with `x::ComponentArray` and the resultant `m.values, m.errors` will be `ComponentArray`s as well.
+- Added `MigradOptimizer()` which can be used as a solver in the `Optimization.jl` ecosystem.
+
 ## 0.4.0 (21-07-2025)
 ### Breaking changes
 - Replaced local distributions (Chebyshev, ArgusBG) with the ones in `DistributionsHEP` package